Convert chart to image in .NET Core
