+++ Charts/Convert-chart-to-image/.NET Framework/Convert-chart-to-image/Data/Template.pptx
