+++ Charts/Convert-chart-to-image/.NET Framework/Convert-chart-to-image/packages.config
@@ -1,0 +1,9 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Syncfusion.Compression.Base" version="20.4.0.38" targetFramework="net461" />
+  <package id="Syncfusion.Licensing" version="20.4.0.38" targetFramework="net461" />
+  <package id="Syncfusion.OfficeChart.Base" version="20.4.0.38" targetFramework="net461" />
+  <package id="Syncfusion.OfficeChartToImageConverter.WinForms" version="20.4.0.38" targetFramework="net461" />
+  <package id="Syncfusion.Presentation.WinForms" version="20.4.0.38" targetFramework="net461" />
+  <package id="Syncfusion.SfChart.WPF" version="20.4.0.38" targetFramework="net461" />
+</packages>
