+++ Charts/Convert-chart-to-image/.NET Standard/Convert-Chart-to-Image.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.4.33213.308
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Convert-Chart-to-Image", "Convert-Chart-to-Image\Convert-Chart-to-Image.csproj", "{85B5A291-A27D-4DCE-8DAB-614EFC130F35}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{85B5A291-A27D-4DCE-8DAB-614EFC130F35}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{85B5A291-A27D-4DCE-8DAB-614EFC130F35}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{85B5A291-A27D-4DCE-8DAB-614EFC130F35}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{85B5A291-A27D-4DCE-8DAB-614EFC130F35}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {2C2E2932-EC8B-41FF-B66C-5534B86E302B}
+	EndGlobalSection
+EndGlobal
