+++ Charts/Convert-chart-to-image/.NET Standard/Convert-Chart-to-Image/Convert-Chart-to-Image.csproj
@@ -1,0 +1,13 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <RootNamespace>Convert_Chart_to_Image</RootNamespace>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Syncfusion.PresentationRenderer.Net.Core" Version="*" />
+  </ItemGroup>
+
+</Project>
